+++ src/compiler/parser/mod.rs
@@ -14,8 +14,10 @@ mod lines_3000_4000;
 pub use lines_3000_4000::*;
 mod lines_4000_5000;
 pub use lines_4000_5000::*;
-mod lines_5000_6000;
-pub use lines_5000_6000::*;
+mod lines_5000_5500;
+pub use lines_5000_5500::*;
+mod lines_5500_6000;
+pub use lines_5500_6000::*;
 mod lines_6000_7000;
 pub use lines_6000_7000::*;
 mod lines_7000_8000;
